requirements.txt
# Add your dependencies here
openai
requests
# Core LangChain dependencies
langchain-core==0.1.41      # ✅ Compatible with langchain-cohere 0.1.0
langchain-cohere==0.1.0
cohere==5.1.4               # ✅ Still required
GitPython==3.1.43
pydantic==2.7.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3